Many bugs!
Fixed the css issue with rendering images on the researchcards, combined 2 functions into 1 to make it clearer when accepting/rejecting research. Added a section to the asm to show the current status of their research and the accept/reject mesage. And made the pell editor on edit profile start with data